Refactor: Split out details panel module
